remove dp_logic.py and query handler
